+++ app/internal/config/__init__.py
@@ -1,0 +1,1 @@
+from .config import RABBITMQ_URL, DESCRIPTION_QUEUE_NAME, PHOTO_QUEUE_NAME
